backend-fastapi/requirements.txt
fastapi
uvicorn
requests
groq
pdfplumber
pymupdf
multipart
python-dotenv
pydantic
psycopg2-binary
supabase
scikit-learn
numpy
pandas
langgraph
langchain_groq
sentence-transformers
beautifulsoup4
google-genai
google-api-python-client
google-auth
httpx

